Fix #32538
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/wt_spamshield/trunk@57086 735d13b6-9817-0410-8766-e36946ffe9aa

+++ ext_tables.sql
@@ -11,7 +11,7 @@ CREATE TABLE tx_wtspamshield_log (
     
     form tinytext NOT NULL,
     errormsg text NOT NULL,
-    pageid tinyint(5) DEFAULT '0' NOT NULL,
+    pageid int(11) DEFAULT '0' NOT NULL,
     formvalues text NOT NULL,
     ip tinytext NOT NULL,
     useragent tinytext NOT NULL,
